Moved User Java files to User package
