+++ app/src/androidTest/java/comcmput301f16t01/github/carrier/LocationTests.java
@@ -5,6 +5,8 @@ import junit.framework.Assert;
 import comcmput301f16t01.github.carrier.Notifications.ElasticNotificationController;
 import comcmput301f16t01.github.carrier.Requests.ElasticRequestController;
 import comcmput301f16t01.github.carrier.Requests.Request;
+import comcmput301f16t01.github.carrier.Users.User;
+import comcmput301f16t01.github.carrier.Users.UserController;
 
 public class LocationTests extends ApplicationTest {
     // University of Alberta, Edmonton
