+++ app/src/androidTest/java/comcmput301f16t01/github/carrier/SearchingTests.java
@@ -10,6 +10,8 @@ import comcmput301f16t01.github.carrier.Requests.ElasticRequestController;
 import comcmput301f16t01.github.carrier.Requests.Request;
 import comcmput301f16t01.github.carrier.Requests.RequestController;
 import comcmput301f16t01.github.carrier.Requests.RequestList;
+import comcmput301f16t01.github.carrier.Users.User;
+import comcmput301f16t01.github.carrier.Users.UserController;
 
 
 public class SearchingTests extends ApplicationTest {
