+++ app/src/androidTest/java/comcmput301f16t01/github/carrier/StatusTests.java
@@ -1,13 +1,11 @@
 package comcmput301f16t01.github.carrier;
 
-import android.location.Location;
-import android.provider.Settings;
-
 import junit.framework.Assert;
 
 import comcmput301f16t01.github.carrier.Requests.ElasticRequestController;
 import comcmput301f16t01.github.carrier.Requests.Request;
 import comcmput301f16t01.github.carrier.Requests.RequestController;
+import comcmput301f16t01.github.carrier.Users.User;
 
 /**
  * Created by meind on 2016-10-11.
