+++ app/src/main/java/comcmput301f16t01/github/carrier/MockObjects/MockNotification.java
@@ -6,7 +6,7 @@ import java.util.Date;
 
 import comcmput301f16t01.github.carrier.Notifications.Notification;
 import comcmput301f16t01.github.carrier.Requests.Request;
-import comcmput301f16t01.github.carrier.User;
+import comcmput301f16t01.github.carrier.Users.User;
 
 /**
  * Mocks a notification. Allows to set the date of a notification instead of it being solely set by
