+++ app/src/main/java/comcmput301f16t01/github/carrier/Notifications/Notification.java
@@ -5,7 +5,7 @@ import android.support.annotation.NonNull;
 import java.util.Date;
 
 import comcmput301f16t01.github.carrier.Requests.Request;
-import comcmput301f16t01.github.carrier.User;
+import comcmput301f16t01.github.carrier.Users.User;
 import io.searchbox.annotations.JestId;
 
 /**
