+++ app/src/main/java/comcmput301f16t01/github/carrier/Requests/Request.java
@@ -1,13 +1,12 @@
 package comcmput301f16t01.github.carrier.Requests;
 
-import android.location.Location;
 import android.support.annotation.NonNull;
 
 import java.util.ArrayList;
 
 //import comcmput301f16t01.github.carrier.Location;
 import comcmput301f16t01.github.carrier.CarrierLocation;
-import comcmput301f16t01.github.carrier.User;
+import comcmput301f16t01.github.carrier.Users.User;
 import io.searchbox.annotations.JestId;
 
 /**
