+++ app/src/main/java/comcmput301f16t01/github/carrier/Requests/RiderRequestActivity.java
@@ -12,8 +12,6 @@ import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.google.gson.Gson;
-
 import org.osmdroid.api.IMapController;
 import org.osmdroid.bonuspack.routing.OSRMRoadManager;
 import org.osmdroid.bonuspack.routing.Road;
@@ -34,11 +32,7 @@ import java.util.Locale;
 
 import comcmput301f16t01.github.carrier.FareCalculator;
 import comcmput301f16t01.github.carrier.R;
-import comcmput301f16t01.github.carrier.Requests.Request;
-import comcmput301f16t01.github.carrier.Requests.RequestController;
-import comcmput301f16t01.github.carrier.User;
-import comcmput301f16t01.github.carrier.UserController;
-import comcmput301f16t01.github.carrier.UsernameTextView;
+import comcmput301f16t01.github.carrier.Users.UsernameTextView;
 
 /**
  * This will help us show the request from the perspective of a rider
