+++ app/src/main/java/comcmput301f16t01/github/carrier/RiderList.java
@@ -2,6 +2,8 @@ package comcmput301f16t01.github.carrier;
 
 import java.util.ArrayList;
 
+import comcmput301f16t01.github.carrier.Users.User;
+
 /**
  * TODO: Remove/delete class when nothing references RiderList
  */
