+++ app/src/main/java/comcmput301f16t01/github/carrier/UserList.java
@@ -2,6 +2,8 @@ package comcmput301f16t01.github.carrier;
 
 import java.util.ArrayList;
 
+import comcmput301f16t01.github.carrier.Users.User;
+
 @Deprecated
 public class UserList extends ArrayList {
