+++ app/src/main/java/comcmput301f16t01/github/carrier/Users/NewUserActivity.java
@@ -1,4 +1,4 @@
-package comcmput301f16t01.github.carrier;
+package comcmput301f16t01.github.carrier.Users;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -7,6 +7,11 @@ import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.Toast;
+
+import comcmput301f16t01.github.carrier.LoginActivity;
+import comcmput301f16t01.github.carrier.LoginMemory;
+import comcmput301f16t01.github.carrier.MainActivity;
+import comcmput301f16t01.github.carrier.R;
 
 /**
  * User can enter their information and a username to create a new account.
