+++ app/src/main/res/layout/activity_driver_view_request.xml
@@ -196,7 +196,7 @@
             android:layout_height="wrap_content"
             android:text="Rider: " />
 
-        <comcmput301f16t01.github.carrier.UsernameTextView
+        <comcmput301f16t01.github.carrier.Users.UsernameTextView
             android:text=""
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -226,7 +226,7 @@
             android:layout_toStartOf="@+id/UsernameTextView_driver"/>
 
 
-        <comcmput301f16t01.github.carrier.UsernameTextView
+        <comcmput301f16t01.github.carrier.Users.UsernameTextView
             android:text=""
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
