+++ app/src/test/java/comcmput301f16t01/github/carrier/RequestControllerTest.java
@@ -1,14 +1,11 @@
 package comcmput301f16t01.github.carrier;
 
-import android.location.Location;
-
 import org.junit.Test;
 
 import comcmput301f16t01.github.carrier.Requests.Request;
 import comcmput301f16t01.github.carrier.Requests.RequestController;
 import comcmput301f16t01.github.carrier.Requests.RequestList;
-
-import static org.junit.Assert.*;
+import comcmput301f16t01.github.carrier.Users.User;
 
 import static org.junit.Assert.*;
