+++ app/src/test/java/comcmput301f16t01/github/carrier/UserControllerTest.java
@@ -2,6 +2,9 @@ package comcmput301f16t01.github.carrier;
 
 import org.junit.Test;
 
+import comcmput301f16t01.github.carrier.Users.User;
+import comcmput301f16t01.github.carrier.Users.UserController;
+
 import static org.junit.Assert.*;
 
 /**
